A more basic example of CDDA playback
This one is useful for testing the bin/cue generation.  

Currently, there is a 75 frame padding added to the .bin, and a pregap on the first track.  This results in a significant audio gap when changing tracks.  On real hardware, it is not there, but playing on emulators is popular so it should probably be fixed.

If the demo is generated as iso/cue/wav and a 2 second postgap and pregap are defined, it will work on both real hardware and emulators, so ideally we'd match the bin/cue generation to match that output somehow.